requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.11.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
ecdsa==0.18.0
fastapi==0.109.1
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.68.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.10
pyasn1==0.5.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.6
sqladmin==0.16.0
SQLAlchemy==2.0.25
starlette==0.35.1
tqdm==4.67.1
typing_extensions==4.9.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.27.0.post1
WTForms==3.1.2
